Add an option to set the `ShareProcessNamespace` property on pods
This seems like it might be useful when using bazel, since it checks
if a server is running by checking a pid file.  Since containers in
pods do not currently share a process namespace, it always appears as
though a new bazel server is needed and drone steps can't take
advantage of the bazel server.